% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdepth.R
\name{edepth}
\alias{edepth}
\title{Compute the extremal depth of a function \code{g} with respect to a set of functions \code{fmat}}
\usage{
edepth(g, fmat, depth_function = "standard")
}
\arguments{
\item{g}{Vector. Function that you want to find the extremal depth of.}

\item{fmat}{Matrix of functions. Each column is a function.}

\item{depth_function}{either "standard" (default) or "rank". Standard orders functions from the center
outward and rank orders least to greatest.}
}
\value{
The extremal depth value as a real number.
}
\description{
Compute the extremal depth of a function \code{g} with respect to a set of functions \code{fmat}
}
