% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdepth.R
\name{edepth_sort}
\alias{edepth_sort}
\title{Compute the extremal depths for all functions in \code{fmat}, with respect to \code{fmat}, then orders them from least to
greatest extremal depth value.}
\usage{
edepth_sort(fmat, depth_function = "standard")
}
\arguments{
\item{fmat}{Matrix of functions. Each column is a function.}

\item{depth_function}{either "standard" (default) or "rank". Standard orders functions from the center
outward and rank orders least to greatest.}
}
\value{
A vector of extremal depth values that correspond to the functions in fmat.
}
\description{
Compute the extremal depths for all functions in \code{fmat}, with respect to \code{fmat}, then orders them from least to
greatest extremal depth value.
}
