% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdepth.R
\name{edepth_set}
\alias{edepth_set}
\title{Compute the extremal depths for all functions in \code{fmat}, with respect to \code{fmat}}
\usage{
edepth_set(fmat)
}
\arguments{
\item{fmat}{Matrix of functions. Each column is a function.}
}
\value{
A vector of extremal depth values that correspond to the functions in fmat.
}
\description{
Compute the extremal depths for all functions in \code{fmat}, with respect to \code{fmat}
}
